文件系统添加业务Id和删除标记

+++ server/file/src/main/java/com/heshaowei/myproj/file/repository/FileBaseInfoRepository.java
@@ -3,5 +3,5 @@ package com.heshaowei.myproj.file.repository;
 import com.heshaowei.myproj.file.entity.FileBaseInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileBaseInfoRepository extends JpaRepository<FileBaseInfo, Integer> {
+public interface FileBaseInfoRepository extends JpaRepository<FileBaseInfo, Long> {
 }
